Refactor code to separate pill component.
