+++ assets/src/edit-story/components/library/panes/media/common/mediaElement.js
@@ -29,7 +29,7 @@ import { rgba } from 'polished';
 import DropDownMenu from '../local/dropDownMenu';
 import { KEYBOARD_USER_SELECTOR } from '../../../../../utils/keyboardOnlyOutline';
 import { useKeyDownEffect } from '../../../../keyboard';
-import useRovingTabIndex from './useRovingTabIndex';
+import { useRovingTabIndex } from '../../shared';
 import Attribution from './attribution';
 import InnerElement from './innerElement';
